今後の予定日を集める責務を移動
RegularEvent自身は今後のイベント日を知っていても違和感はないので、わざわざ別クラスに責任を持たせる必要が無いと思ったため。
